chore: add code documentation.

+++ StockSip.Platform.API/InventoryManagement/Domain/Model/Queries/GetWarehouseByIdQuery.cs
@@ -5,4 +5,4 @@ namespace StockSip.Platform.API.InventoryManagement.Domain.Model.Queries;
 /// This record defines a query to get a warehouse by its unique identifier.
 /// </summary>
 /// <param name="WarehouseId">the unique identifier for the warehouse</param>
-public record GetWarehouseByIdQuery(long WarehouseId);
+public record GetWarehouseByIdQuery(int WarehouseId);
